Add dependency installer
This version allows to install dependencies defined in a JSON formatted
file. Dependencies are installed one by one. Each of them are fetch from
a remote file and installed locally given a path. A checksum can be
optionaly defined to ensure that the proper files are installed. This
will also force them to be re-installed if checksums do not match.

Right now, dependencies are fetched and installed in a linear fashion.
We may improve this by introducing some parallelism in a near future.